feat: add wave pattern to post-card and remove unused components
